+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Parallel functions (part 2):
 
+  - `cancelAll`: Stop all functions running in the background.
   - `cancel`: Stop function running in the background.
   - `wait`: Wait for futures to be completed.
   - `Diary` property for `FevalFuture` object.